raspberry/GatewaySolution/modules/CameraModule/requirements.txt
azure-iot-device~=2.7.0
requests
opencv-python